% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoosounds.R
\name{goToTheZoo}
\alias{goToTheZoo}
\title{goToTheZoo}
\usage{
goToTheZoo(animal, sound)
}
\arguments{
\item{animal}{String The name of an animal}

\item{sound}{String The sound the animal makes}
}
\value{
New string
}
\description{
goToTheZoo
}
